Add rating partial view, Add rating for products

+++ Web/RestaurantSystem.Web.ViewModels/Restaurants/RestaurantInfoViewModel.cs
@@ -25,6 +25,6 @@
 
         public int RatingsCount => this.Ratings.Count();
 
-        public double AverageRating => this.Ratings.Select(x => x.Stars).Sum() / this.RatingsCount;
+        public double AverageRating => (double)this.Ratings.Select(x => x.Stars).Sum() / (double)this.RatingsCount;
     }
 }
